Refactor parser into module so external parser is pluggable
