+++ src/lint/rules/subject-min-length.ts
@@ -1,5 +1,5 @@
 import { RuleCheckResults } from "./";
-import { Commit } from "@commitlint/parse";
+import { Commit } from "lint/parser";
 
 const minLength = 0;
